requirements.txt
torch>=1.13.0
torchaudio>=0.13.0
pytorch-lightning>=1.9.0
transformers>=4.25.1
datasets>=2.7.1
scikit-learn>=1.0.2
pandas>=1.5.2
numpy>=1.23.0
matplotlib>=3.6.0
pyyaml>=6.0
ffmpeg-python>=0.2.0
rich
librosa
soundfile
tqdm
tensorboard
fireducks
jsonargparse[signatures]>=4.27.7